producer/pyproject.toml
[tool.poetry]
name = "kafka-producer"
version = "0.1.0"
description = "Kafka Producer"
authors = ["JMDU"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
boto3 = "^1.26"
numpy = "==1.24.3"
pandas = "==1.5.3"
requests = "^2.28"
kafka-python = "^2.0"

[tool.poetry.dev-dependencies]
black = "^23.1.0"
flake8 = "^6.0.0"
isort = "^5.10.1"

[build-system]
requires = ["poetry-core>=1.2.0"]
build-backend = "poetry.core.masonry.api"
